+++ src/yesoulbike.cpp
@@ -134,7 +134,7 @@ void yesoulbike::characteristicChanged(const QLowEnergyCharacteristic &character
     double distance = GetDistanceFromPacket(newValue);
 
     Resistance = newValue.at(4);
-    Cadence = newValue.at(6);
+    Cadence = ((uint8_t)newValue.at(6));
     m_watts = ((((uint16_t)newValue.at(7)) << 8) + ((uint16_t) newValue.at(8)));
     Speed = 0.37497622 * ((double)Cadence.value());
     KCal += ((( (0.048 * ((double)watts()) + 1.19) * settings.value("weight", 75.0).toFloat() * 3.5) / 200.0 ) / (60000.0 / ((double)lastRefreshCharacteristicChanged.msecsTo(QDateTime::currentDateTime())))); //(( (0.048* Output in watts +1.19) * body weight in kg * 3.5) / 200 ) / 60
